refactor(editor): content query list panel 분리
