+++ front/src/routes/Admin/EditorStudioPage.tsx
@@ -58,6 +58,8 @@ import { EditorStudioSelectedPostToolsPanel } from "./EditorStudioSelectedPostTo
 import { EditorStudioLegacyProfileSection } from "./EditorStudioLegacyProfileSection"
 import { EditorStudioLegacyUtilityPanel } from "./EditorStudioLegacyUtilityPanel"
 import { EditorStudioResultLogPanel } from "./EditorStudioResultLogPanel"
+import { EditorStudioPostQueryPanel } from "./EditorStudioPostQueryPanel"
+import { EditorStudioPostListPanel } from "./EditorStudioPostListPanel"
 import {
   isServerTempDraftPost,
   TEMP_DRAFT_BODY_PLACEHOLDER,
@@ -95,7 +97,6 @@ import {
   readAdminProtectedBootstrap,
 } from "src/libs/server/adminPage"
 import ProfileImage from "src/components/ProfileImage"
-import AppIcon from "src/components/icons/AppIcon"
 import {
   applyThumbnailTransformToUrl,
   clampThumbnailZoom,
@@ -2807,461 +2808,67 @@ export const EditorStudioPage: NextPage<AdminPageProps> = ({ initialMember }) =>
               <ContentStudioLeft
                 data-mobile-visible={!isCompactMobileLayout || activeMobileStudioStep === "query" || activeMobileStudioStep === "list"}
               >
-                <QueryPanel data-mobile-visible={!isCompactMobileLayout || activeMobileStudioStep === "query"}>
-                  <QueryHeader>
-                    <h3>글 목록 조회 조건</h3>
-                    <p>
-                      {listScope === "active"
-                        ? "최근 글을 빠르게 다시 열고 필요한 경우만 고급 조건을 펼쳐 조회합니다."
-                        : "삭제 글만 확인하고 복구 대상을 고릅니다."}
-                    </p>
-                    <ListScopeTabs>
-                      <ListScopeButton
-                        type="button"
-                        data-active={listScope === "active"}
-                        onClick={() => setListScope("active")}
-                      >
-                        활성 글
-                      </ListScopeButton>
-                      <ListScopeButton
-                        type="button"
-                        data-active={listScope === "deleted"}
-                        onClick={() => setListScope("deleted")}
-                      >
-                        삭제 글
-                      </ListScopeButton>
-                    </ListScopeTabs>
-                  </QueryHeader>
-                  <FieldBox>
-                    <FieldLabel htmlFor="list-kw">검색어</FieldLabel>
-                    <Input
-                      id="list-kw"
-                      placeholder={listScope === "active" ? "제목/본문 키워드" : "삭제된 글 제목/본문 키워드"}
-                      value={listKw}
-                      onChange={(e) => setListKw(e.target.value)}
-                    />
-                  </FieldBox>
-
-                  <QueryActions>
-                    <PrimaryButton
-                      disabled={disabled("postList")}
-                      onClick={() => void loadAdminPosts()}
-                    >
-                      {listScope === "active" ? "목록 새로고침" : "삭제 글 조회"}
-                    </PrimaryButton>
-                    {listScope === "active" && (
-                      <Button disabled={disabled("postTemp")} onClick={() => void handleLoadOrCreateTempPost()}>
-                        임시 저장 열기
-                      </Button>
-                    )}
-                  </QueryActions>
-                  {listScope === "active" && (
-                    <PresetRow role="group" aria-label="빠른 프리셋">
-                      <PresetButton
-                        type="button"
-                        data-active={listQuickPreset === "today"}
-                        onClick={() => applyListQuickPreset("today")}
-                      >
-                        오늘 수정
-                      </PresetButton>
-                      <PresetButton
-                        type="button"
-                        data-active={listQuickPreset === "temp"}
-                        onClick={() => applyListQuickPreset("temp")}
-                      >
-                        임시 저장
-                      </PresetButton>
-                      {hasListFiltersApplied && (
-                        <PresetButton
-                          type="button"
-                          data-active={false}
-                          onClick={() => {
-                            setListQuickPreset("none")
-                            setListKw("")
-                            setListPage("1")
-                            setListPageSize("30")
-                            setListSort("CREATED_AT")
-                          }}
-                        >
-                          조건 초기화
-                        </PresetButton>
-                      )}
-                    </PresetRow>
-                  )}
-                  <InlineDisclosure open={isListAdvancedOpen}>
-                    <summary onClick={(event) => {
-                      event.preventDefault()
-                      setIsListAdvancedOpen((prev) => !prev)
-                    }}>
-                      <strong>고급 조회 옵션</strong>
-                      <span>{isListAdvancedOpen ? "닫기" : "열기"}</span>
-                    </summary>
-                    {isListAdvancedOpen && (
-                      <div className="body">
-                        <QueryGrid>
-                          <FieldBox className="listPage">
-                            <FieldLabel htmlFor="list-page">페이지</FieldLabel>
-                            <Input
-                              id="list-page"
-                              type="number"
-                              inputMode="numeric"
-                              min={1}
-                              placeholder="예: 1"
-                              value={listPage}
-                              onChange={handleListPageChange}
-                            />
-                          </FieldBox>
-                          <FieldBox className="listPageSize">
-                            <FieldLabel htmlFor="list-page-size">페이지 크기</FieldLabel>
-                            <Input
-                              id="list-page-size"
-                              type="number"
-                              inputMode="numeric"
-                              min={1}
-                              max={30}
-                              placeholder="1~30"
-                              value={listPageSize}
-                              onChange={handleListPageSizeChange}
-                            />
-                          </FieldBox>
-                          {listScope === "active" && (
-                            <FieldBox className="listSort">
-                              <FieldLabel htmlFor="list-sort">정렬 기준</FieldLabel>
-                              <FieldSelect
-                                id="list-sort"
-                                value={listSort}
-                                onChange={handleListSortChange}
-                              >
-                                {LIST_SORT_OPTIONS.map((option) => (
-                                  <option key={option.value} value={option.value}>
-                                    {option.label}
-                                  </option>
-                                ))}
-                              </FieldSelect>
-                            </FieldBox>
-                          )}
-                        </QueryGrid>
-                      </div>
-                    )}
-                  </InlineDisclosure>
-                </QueryPanel>
-
-              <ListPanel data-mobile-visible={!isCompactMobileLayout || activeMobileStudioStep === "list"}>
-                <ListHeader>
-                  <h3>{listScope === "active" ? "관리자 글 리스트" : "삭제 글 리스트"}</h3>
-                  <ListHeaderActions>
-                    <span>{selectedPostIds.length > 0 ? `${selectedPostIds.length}개 선택` : `총 ${adminPostTotal}건`}</span>
-                    {listScope === "active" ? (
-                      adminPostViewRows.length > 0 ? (
-                        <Button
-                          type="button"
-                          disabled={loadingKey.length > 0}
-                          onClick={toggleSelectAllVisiblePosts}
-                        >
-                          {isAllVisiblePostsSelected ? "현재 목록 선택 해제" : "현재 목록 전체 선택"}
-                        </Button>
-                      ) : null
-                    ) : (
-                      <ReadOnlyHint>삭제 글은 복구 또는 영구삭제로 정리할 수 있습니다.</ReadOnlyHint>
-                    )}
-                  </ListHeaderActions>
-                </ListHeader>
-                {listScope === "active" && selectedPostIds.length > 0 && (
-                  <SelectionStickyBar role="status" aria-live="polite">
-                    <strong>{selectedPostIds.length}개 선택됨</strong>
-                    <div>
-                      <Button type="button" onClick={() => setSelectedPostIds([])} disabled={loadingKey.length > 0}>
-                        선택 해제
-                      </Button>
-                      <Button
-                        type="button"
-                        data-variant="danger"
-                        disabled={loadingKey.length > 0}
-                        onClick={() => openDeleteConfirm(selectedPostIds)}
-                      >
-                        선택 삭제
-                      </Button>
-                    </div>
-                  </SelectionStickyBar>
-                )}
-                {adminPostRows.length === 0 ? (
-                  <ListEmpty>
-                    <p>
-                      {listScope === "active"
-                        ? "목록이 없습니다. 위 조회 조건에서 목록 새로고침을 눌러 시작하세요."
-                        : "삭제된 글이 없습니다. 삭제 글 목록을 조회해 최신 상태를 확인하세요."}
-                    </p>
-                    <div className="actions">
-                      <PrimaryButton
-                        type="button"
-                        disabled={disabled("postList")}
-                        onClick={() => void loadAdminPosts()}
-                      >
-                        {listScope === "active" ? "목록 새로고침" : "삭제 글 목록 조회"}
-                      </PrimaryButton>
-                    </div>
-                  </ListEmpty>
-                ) : (
-                  <>
-                    <ListTableWrap>
-                      <ListTable>
-                      <thead>
-                        <tr>
-                          {listScope === "active" && (
-                            <th className="checkboxCell">
-                              <input
-                                type="checkbox"
-                                aria-label="현재 목록 전체 선택"
-                                checked={isAllVisiblePostsSelected}
-                                onChange={toggleSelectAllVisiblePosts}
-                              />
-                            </th>
-                          )}
-                          <th className="idCell">ID</th>
-                          <th>제목</th>
-                          <th className="dateCell">
-                            {listScope === "active" ? (
-                              <SortHeaderButton
-                                type="button"
-                                onClick={() =>
-                                  setModifiedSortOrder((prev) => (prev === "desc" ? "asc" : "desc"))
-                                }
-                              >
-                                수정일 {modifiedSortOrder === "desc" ? "↓" : "↑"}
-                              </SortHeaderButton>
-                            ) : (
-                              "삭제일"
-                            )}
-                          </th>
-                          <th className="actionsCell">작업</th>
-                        </tr>
-                      </thead>
-                      <tbody>
-                        {adminPostViewRows.map((row) => {
-                          const isLoadedRow = listScope === "active" && editorMode === "edit" && postId.trim() === String(row.id)
-                          return (
-                            <tr key={row.id} data-active={isLoadedRow}>
-                              {listScope === "active" && (
-                                <td className="checkboxCell">
-                                  <input
-                                    type="checkbox"
-                                    aria-label={`${row.id}번 글 선택`}
-                                    checked={selectedPostIdSet.has(row.id)}
-                                    onChange={() => togglePostSelection(row.id)}
-                                  />
-                                </td>
-                              )}
-                              <td className="idCell">{row.id}</td>
-                              <td className="title">
-                                <TitleCell>
-                                  <div className="titleMain">
-                                    <span className="text">{row.title}</span>
-                                    {isLoadedRow && <LoadedBadge>현재 편집 중</LoadedBadge>}
-                                    {listScope === "deleted" && <DeletedBadge>삭제됨</DeletedBadge>}
-                                    <VisibilityBadge className="inlineVisibility" data-tone={toVisibility(row.published, row.listed)}>
-                                      {getVisibilityLabel(row.published, row.listed)}
-                                    </VisibilityBadge>
-                                  </div>
-                                  <span className="meta">{row.authorName || "작성자 미상"}</span>
-                                </TitleCell>
-                              </td>
-                              <td className="dateCell">{(listScope === "deleted" ? row.deletedAt : row.modifiedAt)?.slice(0, 10) || "-"}</td>
-                              <td className="actionsCell">
-                                <InlineActions>
-                                  {listScope === "active" ? (
-                                    <>
-                                      <RowActionButton
-                                        type="button"
-                                        data-variant="primary"
-                                        disabled={loadingKey.length > 0}
-                                        onClick={() => {
-                                          setPostId(String(row.id))
-                                          void loadPostForEditor(String(row.id))
-                                        }}
-                                      >
-                                        <AppIcon name="edit" />
-                                        <span>{isLoadedRow ? "계속 편집" : "편집"}</span>
-                                      </RowActionButton>
-                                      <RowActionMenu>
-                                        <summary>
-                                          <span>더보기</span>
-                                          <AppIcon name="chevron-down" />
-                                        </summary>
-                                        <div className="menu">
-                                          <button
-                                            type="button"
-                                            disabled={loadingKey.length > 0}
-                                            onClick={() => void openPostDetailRoute(row.id)}
-                                          >
-                                            상세 열기
-                                          </button>
-                                          <button
-                                            type="button"
-                                            disabled={loadingKey.length > 0}
-                                            onClick={() => void copyPostDetailLink(row.id, row.title)}
-                                          >
-                                            링크 복사
-                                          </button>
-                                          <button
-                                            type="button"
-                                            disabled={loadingKey.length > 0}
-                                            onClick={() => openDeleteConfirm([row.id], row.title)}
-                                          >
-                                            삭제
-                                          </button>
-                                        </div>
-                                      </RowActionMenu>
-                                    </>
-                                  ) : (
-                                    <>
-                                      <RowActionButton
-                                        type="button"
-                                        data-variant="primary"
-                                        disabled={loadingKey.length > 0}
-                                        onClick={() => void restoreDeletedPostFromList(row)}
-                                      >
-                                        <AppIcon name="check-circle" />
-                                        <span>복구</span>
-                                      </RowActionButton>
-                                      <RowActionMenu>
-                                        <summary>
-                                          <span>더보기</span>
-                                          <AppIcon name="chevron-down" />
-                                        </summary>
-                                        <div className="menu">
-                                          <button
-                                            type="button"
-                                            disabled={loadingKey.length > 0}
-                                            onClick={() => void hardDeleteDeletedPostFromList(row)}
-                                          >
-                                            영구삭제
-                                          </button>
-                                        </div>
-                                      </RowActionMenu>
-                                    </>
-                                  )}
-                                </InlineActions>
-                              </td>
-                            </tr>
-                          )
-                        })}
-                      </tbody>
-                      </ListTable>
-                    </ListTableWrap>
-
-                    <MobileListCards>
-                      {adminPostViewRows.map((row) => {
-                        const isLoadedRow = listScope === "active" && editorMode === "edit" && postId.trim() === String(row.id)
-                        return (
-                        <article key={`mobile-${row.id}`} data-active={isLoadedRow}>
-                          <header>
-                            <div className="metaLeading">
-                              {listScope === "active" && (
-                                <input
-                                  type="checkbox"
-                                  aria-label={`${row.id}번 글 선택`}
-                                  checked={selectedPostIdSet.has(row.id)}
-                                  onChange={() => togglePostSelection(row.id)}
-                                />
-                              )}
-                              <span className="rowId">#{row.id}</span>
-                            </div>
-                            {isLoadedRow ? <LoadedBadge>현재 편집 중</LoadedBadge> : null}
-                          </header>
-                          <h4>{row.title}</h4>
-                          <p className="metaLine">
-                            <span>
-                              {row.authorName}
-                              <span className="dot">•</span>
-                              {(listScope === "deleted" ? row.deletedAt : row.modifiedAt)?.slice(0, 10) || "-"}
-                            </span>
-                            <VisibilityBadge data-tone={toVisibility(row.published, row.listed)}>
-                              {getVisibilityLabel(row.published, row.listed)}
-                            </VisibilityBadge>
-                          </p>
-                          <div className="mainAction">
-                            {listScope === "active" ? (
-                              <>
-                                <RowActionButton
-                                  type="button"
-                                  data-variant="primary"
-                                  disabled={loadingKey.length > 0}
-                                  onClick={() => {
-                                    setPostId(String(row.id))
-                                    void loadPostForEditor(String(row.id))
-                                  }}
-                                >
-                                  <AppIcon name="edit" />
-                                  <span>{isLoadedRow ? "계속 편집" : "편집"}</span>
-                                </RowActionButton>
-                                <RowActionMenu>
-                                  <summary>
-                                    <span>더보기</span>
-                                    <AppIcon name="chevron-down" />
-                                  </summary>
-                                  <div className="menu">
-                                    <button
-                                      type="button"
-                                      disabled={loadingKey.length > 0}
-                                      onClick={() => void openPostDetailRoute(row.id)}
-                                    >
-                                      상세 열기
-                                    </button>
-                                    <button
-                                      type="button"
-                                      disabled={loadingKey.length > 0}
-                                      onClick={() => void copyPostDetailLink(row.id, row.title)}
-                                    >
-                                      링크 복사
-                                    </button>
-                                    <button
-                                      type="button"
-                                      disabled={loadingKey.length > 0}
-                                      onClick={() => openDeleteConfirm([row.id], row.title)}
-                                    >
-                                      삭제
-                                    </button>
-                                  </div>
-                                </RowActionMenu>
-                              </>
-                            ) : (
-                              <>
-                                <RowActionButton
-                                  type="button"
-                                  data-variant="primary"
-                                  disabled={loadingKey.length > 0}
-                                  onClick={() => void restoreDeletedPostFromList(row)}
-                                >
-                                  <AppIcon name="check-circle" />
-                                  <span>복구</span>
-                                </RowActionButton>
-                                <RowActionMenu>
-                                  <summary>
-                                    <span>더보기</span>
-                                    <AppIcon name="chevron-down" />
-                                  </summary>
-                                  <div className="menu">
-                                    <button
-                                      type="button"
-                                      disabled={loadingKey.length > 0}
-                                      onClick={() => void hardDeleteDeletedPostFromList(row)}
-                                    >
-                                      영구삭제
-                                    </button>
-                                  </div>
-                                </RowActionMenu>
-                              </>
-                            )}
-                          </div>
-                        </article>
-                      )})}
-                    </MobileListCards>
-                  </>
-                )}
-                {listScope === "deleted" && deletedListNotice.text && (
-                  <InlineStatus data-tone={deletedListNotice.tone}>{deletedListNotice.text}</InlineStatus>
-                )}
-              </ListPanel>
+                <EditorStudioPostQueryPanel
+                  activeMobileStep={activeMobileStudioStep as ManageMobileStudioStep}
+                  isCompactMobileLayout={isCompactMobileLayout}
+                  listScope={listScope}
+                  listKeyword={listKw}
+                  listQuickPreset={listQuickPreset}
+                  hasListFiltersApplied={hasListFiltersApplied}
+                  isAdvancedOpen={isListAdvancedOpen}
+                  listPage={listPage}
+                  listPageSize={listPageSize}
+                  listSort={listSort}
+                  listSortOptions={LIST_SORT_OPTIONS}
+                  isRefreshDisabled={disabled("postList")}
+                  isTempPostDisabled={disabled("postTemp")}
+                  onListScopeChange={setListScope}
+                  onListKeywordChange={setListKw}
+                  onRefreshList={() => void loadAdminPosts()}
+                  onLoadOrCreateTempPost={() => void handleLoadOrCreateTempPost()}
+                  onApplyQuickPreset={applyListQuickPreset}
+                  onResetFilters={() => {
+                    setListQuickPreset("none")
+                    setListKw("")
+                    setListPage("1")
+                    setListPageSize("30")
+                    setListSort("CREATED_AT")
+                  }}
+                  onToggleAdvanced={() => setIsListAdvancedOpen((prev) => !prev)}
+                  onListPageChange={handleListPageChange}
+                  onListPageSizeChange={handleListPageSizeChange}
+                  onListSortChange={handleListSortChange}
+                />
+                <EditorStudioPostListPanel
+                  mobileVisible={!isCompactMobileLayout || activeMobileStudioStep === "list"}
+                  listScope={listScope}
+                  selectedPostIds={selectedPostIds}
+                  adminPostTotal={adminPostTotal}
+                  adminPostRows={adminPostRows}
+                  adminPostViewRows={adminPostViewRows}
+                  isAllVisiblePostsSelected={isAllVisiblePostsSelected}
+                  selectedPostIdSet={selectedPostIdSet}
+                  editorMode={editorMode}
+                  postId={postId}
+                  loadingKey={loadingKey}
+                  modifiedSortOrder={modifiedSortOrder}
+                  deletedListNotice={deletedListNotice}
+                  isRefreshDisabled={disabled("postList")}
+                  onToggleSelectAllVisiblePosts={toggleSelectAllVisiblePosts}
+                  onClearSelection={() => setSelectedPostIds([])}
+                  onRequestDeletePosts={(ids, headline) => openDeleteConfirm(ids, headline)}
+                  onRefreshList={() => void loadAdminPosts()}
+                  onTogglePostSelection={togglePostSelection}
+                  onToggleModifiedSortOrder={() => setModifiedSortOrder((prev) => (prev === "desc" ? "asc" : "desc"))}
+                  onEditPost={(row) => {
+                    setPostId(String(row.id))
+                    void loadPostForEditor(String(row.id))
+                  }}
+                  onOpenPostDetail={(id) => void openPostDetailRoute(id)}
+                  onCopyPostDetailLink={(id, title) => void copyPostDetailLink(id, title)}
+                  onRestoreDeletedPost={(row) => void restoreDeletedPostFromList(row)}
+                  onHardDeletePost={(row) => void hardDeleteDeletedPostFromList(row)}
+                />
               </ContentStudioLeft>
 
               <EditorStudioSelectedPostPanel
@@ -4041,95 +3648,6 @@ const ContentStudioLeft = styled.div`
   }
 `
 
-const QueryPanel = styled.div`
-  border: 1px solid ${({ theme }) => theme.colors.gray6};
-  border-radius: 10px;
-  background: ${({ theme }) => theme.colors.gray2};
-  padding: 0.72rem 0.72rem 0.82rem;
-  margin: 0;
-  border-bottom: 1px solid ${({ theme }) => theme.colors.gray6};
-
-  @media (max-width: 420px) {
-    padding: 0.62rem 0.62rem 0.72rem;
-  }
-
-  @media (max-width: 720px) {
-    &[data-mobile-visible="false"] {
-      display: none;
-    }
-  }
-`
-
-const QueryHeader = styled.div`
-  margin-bottom: 0.72rem;
-
-  h3 {
-    margin: 0;
-    font-size: 1rem;
-    font-weight: 720;
-    color: ${({ theme }) => theme.colors.gray12};
-  }
-
-  p {
-    margin: 0.18rem 0 0;
-    font-size: 0.78rem;
-    line-height: 1.5;
-    color: ${({ theme }) => theme.colors.gray10};
-  }
-`
-
-const ListScopeTabs = styled.div`
-  display: inline-flex;
-  align-items: center;
-  gap: 0.38rem;
-  margin-top: 0.55rem;
-  padding: 0;
-  border-radius: 999px;
-  border: none;
-  background: transparent;
-`
-
-const ListScopeButton = styled.button`
-  border: 0;
-  border-radius: 999px;
-  min-height: 36px;
-  padding: 0 0.82rem;
-  background: transparent;
-  color: ${({ theme }) => theme.colors.gray11};
-  font-size: 0.82rem;
-  font-weight: 700;
-  cursor: pointer;
-  transition:
-    background 0.18s ease,
-    color 0.18s ease;
-
-  &[data-active="true"] {
-    background: transparent;
-    color: ${({ theme }) => theme.colors.gray12};
-    text-decoration: underline;
-    text-underline-offset: 4px;
-  }
-`
-
-const QueryGrid = styled.div`
-  display: grid;
-  grid-template-columns: repeat(2, minmax(0, 1fr));
-  gap: 0.5rem;
-
-  @media (min-width: 1280px) {
-    grid-template-columns: 110px 140px minmax(260px, 1fr) 180px;
-    align-items: end;
-
-    .listKw {
-      min-width: 0;
-    }
-  }
-
-  @media (max-width: 720px) {
-    grid-template-columns: 1fr;
-  }
-`
-
 const FieldBox = styled.div`
   display: grid;
   gap: 0.26rem;
@@ -4147,105 +3665,6 @@ const FieldLabel = styled.label`
   font-size: 0.8rem;
   font-weight: 650;
   color: ${({ theme }) => theme.colors.gray11};
-`
-
-const QueryActions = styled.div`
-  margin-top: 0.72rem;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-
-  > button {
-    min-width: 9.5rem;
-  }
-
-  @media (max-width: 720px) {
-    display: grid;
-    grid-template-columns: 1fr;
-
-    > button {
-      width: 100%;
-    }
-  }
-`
-
-const InlineDisclosure = styled.details`
-  margin-top: 0.68rem;
-  border-top: 1px dashed ${({ theme }) => theme.colors.gray6};
-  padding-top: 0.68rem;
-
-  summary {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 0.72rem;
-    list-style: none;
-    cursor: pointer;
-    color: ${({ theme }) => theme.colors.gray11};
-    font-size: 0.78rem;
-    line-height: 1.45;
-
-    &::-webkit-details-marker {
-      display: none;
-    }
-  }
-
-  strong {
-    color: ${({ theme }) => theme.colors.gray12};
-    font-size: 0.8rem;
-    font-weight: 700;
-  }
-
-  summary > span:last-of-type {
-    flex: 0 0 auto;
-    color: ${({ theme }) => theme.colors.blue11};
-    font-size: 0.76rem;
-    font-weight: 700;
-  }
-
-  .body {
-    display: grid;
-    gap: 0.62rem;
-    margin-top: 0.72rem;
-  }
-`
-
-const PresetRow = styled.div`
-  margin-top: 0.6rem;
-  display: flex;
-  gap: 0.4rem;
-  flex-wrap: wrap;
-
-  @media (max-width: 420px) {
-    flex-wrap: nowrap;
-    overflow-x: auto;
-    padding-bottom: 0.18rem;
-    scrollbar-width: none;
-    -webkit-overflow-scrolling: touch;
-
-    &::-webkit-scrollbar {
-      display: none;
-    }
-  }
-`
-
-const PresetButton = styled.button`
-  min-height: 36px;
-  border-radius: 999px;
-  border: 1px solid ${({ theme }) => theme.colors.gray6};
-  background: transparent;
-  color: ${({ theme }) => theme.colors.gray10};
-  font-size: 0.78rem;
-  font-weight: 700;
-  padding: 0 0.72rem;
-  cursor: pointer;
-
-  &[data-active="true"] {
-    color: ${({ theme }) => theme.colors.gray12};
-    border-color: ${({ theme }) => theme.colors.gray7};
-    text-decoration: underline;
-    text-underline-offset: 4px;
-  }
 `
 
 const InlineStatus = styled.div`
@@ -4326,23 +3745,6 @@ const Input = styled.input`
   min-width: 0;
   background: transparent;
   color: ${({ theme }) => theme.colors.gray12};
-
-  &:focus-visible {
-    outline: none;
-    border-color: ${({ theme }) => theme.colors.blue8};
-    box-shadow: 0 0 0 4px ${({ theme }) => theme.colors.blue4};
-  }
-`
-
-const FieldSelect = styled.select`
-  border: 1px solid ${({ theme }) => theme.colors.gray6};
-  border-radius: 8px;
-  padding: 0.72rem 0.8rem;
-  min-height: 44px;
-  min-width: 0;
-  background: ${({ theme }) => theme.colors.gray1};
-  color: ${({ theme }) => theme.colors.gray12};
-  font-size: 0.95rem;
 
   &:focus-visible {
     outline: none;
@@ -4881,140 +4283,6 @@ const SelectedTagChip = styled.span`
   }
 `
 
-const ListPanel = styled.div`
-  border: 1px solid ${({ theme }) => theme.colors.gray5};
-  border-radius: 12px;
-  background: ${({ theme }) => theme.colors.gray1};
-  padding: 0.82rem;
-  margin: 0;
-  min-width: 0;
-  display: grid;
-  gap: 0.62rem;
-
-  @media (max-width: 720px) {
-    &[data-mobile-visible="false"] {
-      display: none;
-    }
-  }
-`
-
-const ListHeader = styled.div`
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  gap: 0.64rem;
-  margin-bottom: 0.75rem;
-
-  h3 {
-    margin: 0;
-    font-size: 1rem;
-    font-weight: 720;
-    color: ${({ theme }) => theme.colors.gray12};
-  }
-
-  span {
-    font-size: 0.8rem;
-    color: ${({ theme }) => theme.colors.gray11};
-  }
-
-  @media (max-width: 920px) {
-    flex-direction: column;
-  }
-`
-
-const ListHeaderActions = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-  justify-content: flex-end;
-  align-items: center;
-
-  span {
-    display: inline-flex;
-    align-items: center;
-    min-height: 34px;
-    padding: 0 0.72rem;
-    border-radius: 999px;
-    border: 1px solid ${({ theme }) => theme.colors.gray6};
-    background: ${({ theme }) => theme.colors.gray2};
-    color: ${({ theme }) => theme.colors.gray12};
-    font-size: 0.8rem;
-    font-weight: 700;
-    white-space: nowrap;
-  }
-
-  @media (max-width: 920px) {
-    justify-content: flex-start;
-  }
-
-  @media (max-width: 720px) {
-    width: 100%;
-
-    span {
-      width: 100%;
-      margin-right: 0;
-    }
-
-    > button {
-      width: 100%;
-      justify-content: center;
-    }
-  }
-`
-
-const ReadOnlyHint = styled.span`
-  display: inline-flex;
-  align-items: center;
-  border-radius: 999px;
-  border: 1px solid ${({ theme }) => theme.colors.gray6};
-  background: ${({ theme }) => theme.colors.gray2};
-  color: ${({ theme }) => theme.colors.gray11};
-  min-height: 28px;
-  padding: 0 0.58rem;
-  font-size: 0.72rem;
-  font-weight: 600;
-`
-
-const ListEmpty = styled.div`
-  margin: 0;
-  min-height: 13.5rem;
-  display: grid;
-  place-items: center;
-  text-align: center;
-  padding: 0.8rem 1rem;
-  border-radius: 10px;
-  border: 1px dashed ${({ theme }) => theme.colors.gray6};
-  color: ${({ theme }) => theme.colors.gray11};
-  gap: 0.72rem;
-
-  p {
-    margin: 0;
-    font-size: 0.86rem;
-    line-height: 1.65;
-  }
-
-  .actions {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    gap: 0.5rem;
-  }
-`
-
-const ListTableWrap = styled.div`
-  width: 100%;
-  overflow-x: auto;
-  overflow-y: auto;
-  max-height: 52vh;
-  border: 1px solid ${({ theme }) => theme.colors.gray6};
-  border-radius: 10px;
-  overscroll-behavior: contain;
-
-  @media (max-width: 1100px) {
-    display: none;
-  }
-`
-
 const SubActionRow = styled.div`
   display: flex;
   flex-wrap: wrap;
@@ -5034,467 +4302,6 @@ const SubActionRow = styled.div`
     > button {
       width: 100%;
       justify-content: center;
-    }
-  }
-`
-
-const SelectionStickyBar = styled.div`
-  position: sticky;
-  top: 0;
-  z-index: 2;
-  margin: 0 0 0.68rem;
-  padding: 0.55rem 0.62rem;
-  border: 1px solid ${({ theme }) => theme.colors.gray6};
-  border-radius: 10px;
-  background: ${({ theme }) => theme.colors.gray2};
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 0.5rem;
-  flex-wrap: wrap;
-
-  strong {
-    color: ${({ theme }) => theme.colors.gray12};
-    font-size: 0.8rem;
-  }
-
-  > div {
-    display: flex;
-    gap: 0.4rem;
-    flex-wrap: wrap;
-  }
-
-  @media (max-width: 900px) {
-    top: calc(var(--app-header-height, 56px) + 0.35rem);
-  }
-`
-
-const ListTable = styled.table`
-  width: 100%;
-  min-width: 980px;
-  border-collapse: collapse;
-  table-layout: fixed;
-
-  th,
-  td {
-    border-bottom: 1px solid ${({ theme }) => theme.colors.gray6};
-    padding: 0.5rem 0.45rem;
-    text-align: left;
-    font-size: 0.8rem;
-    color: ${({ theme }) => theme.colors.gray12};
-    vertical-align: middle;
-  }
-
-  th {
-    position: sticky;
-    top: 0;
-    z-index: 1;
-    background: ${({ theme }) => theme.colors.gray2};
-    font-size: 0.75rem;
-    color: ${({ theme }) => theme.colors.gray11};
-    font-weight: 700;
-  }
-
-  tbody tr:last-of-type td {
-    border-bottom: 0;
-  }
-
-  tbody tr {
-    transition: background-color 0.18s ease, box-shadow 0.18s ease;
-  }
-
-  tbody tr:hover td {
-    background: rgba(255, 255, 255, 0.02);
-  }
-
-  tbody tr[data-active="true"] td {
-    background:
-      linear-gradient(90deg, rgba(59, 130, 246, 0.14) 0, rgba(59, 130, 246, 0.04) 28px, rgba(255, 255, 255, 0.02) 28px);
-  }
-
-  .checkboxCell {
-    width: 2rem;
-    text-align: center;
-    padding-left: 0.2rem;
-    padding-right: 0.2rem;
-  }
-
-  th.idCell,
-  td.idCell {
-    width: 4.75rem;
-    white-space: nowrap;
-  }
-
-  input[type="checkbox"] {
-    width: 0.92rem;
-    height: 0.92rem;
-    cursor: pointer;
-    accent-color: ${({ theme }) => theme.colors.blue9};
-  }
-
-  td.title {
-    min-width: 0;
-  }
-
-  th.dateCell,
-  td.dateCell {
-    width: 112px;
-    white-space: nowrap;
-  }
-
-  th.actionsCell,
-  td.actionsCell {
-    width: 132px;
-    min-width: 132px;
-  }
-
-  @media (max-width: 1520px) {
-    th.actionsCell,
-    td.actionsCell {
-      width: 124px;
-      min-width: 124px;
-    }
-  }
-`
-
-const TitleCell = styled.div`
-  display: grid;
-  gap: 0.36rem;
-  max-width: 100%;
-  min-width: 0;
-
-  .titleMain {
-    display: flex;
-    align-items: center;
-    gap: 0.4rem;
-    min-width: 0;
-    flex-wrap: wrap;
-  }
-
-  .text {
-    min-width: 0;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    color: ${({ theme }) => theme.colors.gray12};
-    font-weight: 700;
-  }
-
-  .meta {
-    display: inline-flex;
-    align-items: center;
-    min-width: 0;
-    color: ${({ theme }) => theme.colors.gray10};
-    font-size: 0.72rem;
-    font-weight: 600;
-    white-space: nowrap;
-  }
-
-  .inlineVisibility {
-    display: inline-flex;
-  }
-`
-
-const DeletedBadge = styled.span`
-  display: inline-flex;
-  align-items: center;
-  border-radius: 999px;
-  border: 1px solid ${({ theme }) => theme.colors.gray6};
-  background: transparent;
-  color: ${({ theme }) => theme.colors.gray11};
-  font-size: 0.68rem;
-  font-weight: 700;
-  padding: 0.12rem 0.42rem;
-  flex: 0 0 auto;
-`
-
-const LoadedBadge = styled.span`
-  display: inline-flex;
-  align-items: center;
-  min-height: 24px;
-  border-radius: 999px;
-  border: 1px solid ${({ theme }) => theme.colors.blue7};
-  background: ${({ theme }) => theme.colors.blue3};
-  color: ${({ theme }) => theme.colors.blue11};
-  padding: 0 0.5rem;
-  font-size: 0.7rem;
-  font-weight: 800;
-  line-height: 1;
-  flex: 0 0 auto;
-`
-
-const SortHeaderButton = styled.button`
-  border: 0;
-  background: transparent;
-  padding: 0;
-  color: ${({ theme }) => theme.colors.gray11};
-  font-size: 0.74rem;
-  font-weight: 700;
-  cursor: pointer;
-`
-
-const VisibilityBadge = styled.span`
-  display: inline-flex;
-  align-items: center;
-  border-radius: 999px;
-  padding: 0.16rem 0.46rem;
-  font-size: 0.72rem;
-  border: 1px solid ${({ theme }) => theme.colors.gray6};
-  color: ${({ theme }) => theme.colors.gray11};
-  background: ${({ theme }) => theme.colors.gray2};
-
-  &[data-tone="PRIVATE"] {
-    color: ${({ theme }) => theme.colors.gray11};
-  }
-
-  &[data-tone="PUBLIC_UNLISTED"] {
-    color: ${({ theme }) => theme.colors.blue11};
-    border-color: ${({ theme }) => theme.colors.blue7};
-    background: ${({ theme }) => theme.colors.blue3};
-  }
-
-  &[data-tone="PUBLIC_LISTED"] {
-    color: ${({ theme }) => theme.colors.green11};
-    border-color: ${({ theme }) => theme.colors.green7};
-    background: ${({ theme }) => theme.colors.green3};
-  }
-`
-
-const InlineActions = styled.div`
-  display: grid;
-  gap: 0.42rem;
-  align-items: stretch;
-`
-
-const RowActionButton = styled(Button)`
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.42rem;
-  width: 100%;
-  min-height: 40px;
-  padding: 0.42rem 0.62rem;
-  font-size: 0.78rem;
-  font-weight: 700;
-  white-space: nowrap;
-
-  svg {
-    flex: 0 0 auto;
-  }
-
-  span {
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-
-  &[data-variant="primary"] {
-    border-color: ${({ theme }) => theme.colors.blue8};
-    background: ${({ theme }) => theme.colors.blue3};
-    color: ${({ theme }) => theme.colors.blue11};
-  }
-
-  &[data-variant="primary"]:hover:not(:disabled) {
-    border-color: ${({ theme }) => theme.colors.blue9};
-    background: ${({ theme }) => theme.colors.blue4};
-    color: ${({ theme }) => theme.colors.blue12};
-  }
-
-  &[data-variant="soft-danger"] {
-    border-color: rgba(239, 68, 68, 0.38);
-    background: rgba(127, 29, 29, 0.16);
-    color: ${({ theme }) => theme.colors.red11};
-  }
-
-  &[data-variant="soft-danger"]:hover:not(:disabled) {
-    border-color: ${({ theme }) => theme.colors.red8};
-    background: ${({ theme }) => theme.colors.red3};
-    color: ${({ theme }) => theme.colors.red11};
-  }
-
-  &[data-variant="subtle-danger"] {
-    border-color: transparent;
-    background: transparent;
-    color: ${({ theme }) => theme.colors.red10};
-  }
-
-  &[data-variant="subtle-danger"]:hover:not(:disabled) {
-    border-color: rgba(239, 68, 68, 0.22);
-    background: rgba(127, 29, 29, 0.12);
-    color: ${({ theme }) => theme.colors.red11};
-  }
-`
-
-const RowActionMenu = styled.details`
-  position: relative;
-
-  summary {
-    list-style: none;
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    gap: 0.3rem;
-    width: 100%;
-    min-height: 40px;
-    border-radius: 8px;
-    border: 1px solid ${({ theme }) => theme.colors.gray6};
-    background: transparent;
-    color: ${({ theme }) => theme.colors.gray11};
-    font-size: 0.76rem;
-    font-weight: 700;
-    cursor: pointer;
-
-    &::-webkit-details-marker {
-      display: none;
-    }
-  }
-
-  .menu {
-    position: absolute;
-    right: 0;
-    top: calc(100% + 0.3rem);
-    z-index: 8;
-    display: grid;
-    min-width: 7rem;
-    padding: 0.32rem;
-    border-radius: 10px;
-    border: 1px solid ${({ theme }) => theme.colors.gray6};
-    background: ${({ theme }) => theme.colors.gray2};
-    box-shadow: none;
-  }
-
-  .menu button {
-    min-height: 36px;
-    border: 0;
-    border-radius: 8px;
-    background: transparent;
-    color: ${({ theme }) => theme.colors.red11};
-    font-size: 0.78rem;
-    font-weight: 700;
-    cursor: pointer;
-  }
-
-  .menu button:hover:not(:disabled) {
-    background: ${({ theme }) => theme.colors.red3};
-  }
-
-  .menu button:disabled {
-    opacity: 0.45;
-    cursor: not-allowed;
-  }
-`
-
-const MobileListCards = styled.div`
-  display: none;
-  margin-top: 0.65rem;
-
-  @media (max-width: 1100px) {
-    display: grid;
-    gap: 0.6rem;
-  }
-
-  article {
-    border: 1px solid ${({ theme }) => theme.colors.gray6};
-    border-radius: 10px;
-    padding: 0.62rem;
-    background: ${({ theme }) => theme.colors.gray2};
-    display: grid;
-    gap: 0.5rem;
-    content-visibility: auto;
-    contain-intrinsic-size: 1px 172px;
-    transition: background-color 0.18s ease, border-color 0.18s ease, box-shadow 0.18s ease;
-  }
-
-  article[data-active="true"] {
-    border-color: ${({ theme }) => theme.colors.blue7};
-    background: ${({ theme }) => theme.colors.blue3};
-    box-shadow: none;
-  }
-
-  header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 0.45rem;
-  }
-
-  .metaLeading {
-    display: inline-flex;
-    align-items: center;
-    gap: 0.4rem;
-    min-width: 0;
-  }
-
-  .metaLeading input[type="checkbox"] {
-    width: 1rem;
-    height: 1rem;
-    accent-color: ${({ theme }) => theme.colors.gray10};
-  }
-
-  h4 {
-    margin: 0;
-    color: ${({ theme }) => theme.colors.gray12};
-    font-size: 0.9rem;
-    line-height: 1.45;
-    word-break: break-word;
-    display: -webkit-box;
-    -webkit-line-clamp: 2;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-  }
-
-  p {
-    margin: 0;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 0.4rem;
-    color: ${({ theme }) => theme.colors.gray10};
-    font-size: 0.78rem;
-  }
-
-  .metaLine {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 0.42rem;
-
-    .dot {
-      margin: 0 0.26rem;
-      opacity: 0.65;
-    }
-  }
-
-  .mainAction {
-    display: grid;
-    grid-template-columns: repeat(2, minmax(0, 1fr));
-    gap: 0.42rem;
-  }
-
-  .mainAction > button {
-    width: 100%;
-    justify-content: center;
-  }
-
-  .rowId {
-    color: ${({ theme }) => theme.colors.gray10};
-    font-size: 0.78rem;
-  }
-
-  @media (max-width: 420px) {
-    gap: 0.52rem;
-
-    article {
-      padding: 0.56rem;
-    }
-
-    p {
-      align-items: flex-start;
-      flex-wrap: wrap;
-      justify-content: flex-start;
-    }
-
-    .mainAction {
-      grid-template-columns: 1fr;
     }
   }
 `
